\name{cm.zavaleta}
\docType{data}
\alias{cm.zavaleta}
\title{A Seven-Variable Lyme Disease Vector—Host Community}
\description{
This dataset represents the causal relationships amongst the seven variables defined by Zavaleta's 2004 article in figure 1, page 613.
}
\usage{data(cm.zavaleta)}
\format{A 7 by 7 (-1, 0, 1) matrix with each element having at least one parent and at least one child.  The matrix is interpreted as the jth variable has the direct causal effect a[i,j] on variable i.}
\references{
  Puccia, C. J. and Levins, R. (1986) \emph{Qualitative Modeling of Complex Systems: An Introduction to Loop Analysis and Time Averaging}. Cambridge: Harvard University Press.

  Zavaleta, J. and Rossignol, P. (2004) Community-level analysis of risk of vector-borne disease. \emph{Transactions of the Royal Society of Tropical Medicine and Hygiene}, \bold{98(10)},610-618.
}
\keyword{datasets}
