\name{community.effect.matrix}
\docType{data}
\alias{community.effect.matrix}
\title{The Community Effect Matrix of a Simulated Five-Parameter Data Set}
\description{
This dataset contains the computed community effect matrix of a five-parameter system.
}
\usage{data(community.effect.matrix)}
\format{A 5 by 5 (-1, 0, 1, NA) matrix.  The matrix is interpreted as the ith parameter has the indirect causal effect [i,j] on parameter j after the system stabilizes in response to a perturbation.}
\references{
  Puccia, C. J. and Levins, R. (1986) \emph{Qualitative Modeling of Complex Systems: An Introduction to Loop Analysis and Time Averaging}. Cambridge: Harvard University Press.
}
\keyword{datasets}
