\name{community.matrix}
\docType{data}
\alias{community.matrix}
\title{Simulated Five-Variable Data Set}
\description{
This dataset represents the causal relationships defining a five-variable system.
}
\usage{data(community.matrix)}
\format{A 5 by 5 (-1, 0, 1) matrix with each element having at least one parent and at least one child.  The matrix is interpreted as the jth variable has the direct causal effect a[i,j] on variable i.}
\references{
  Puccia, C. J. and Levins, R. (1986) \emph{Qualitative Modeling of Complex Systems: An Introduction to Loop Analysis and Time Averaging}. Cambridge: Harvard University Press.
}
\keyword{datasets}
