\name{LoopAnalyst-package}
\alias{LoopAnalyst-package}
\alias{LoopAnalyst}
\docType{package}
\title{
\packageTitle{LoopAnalyst}
}
\description{
\packageDescription{LoopAnalyst}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LoopAnalyst}
\packageIndices{LoopAnalyst}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{LoopAnalyst}

Maintainer: \packageMaintainer{LoopAnalyst}
}
\references{
 Dambacher, J. M. and Li, H. W. and Rossignol, P. A. (2002) \href{https://www.jstor.org/stable/3071950?seq=1#page_scan_tab_contents}{Relevance of community structure in assessing indeterminacy of ecological predictions}. \emph{Ecology}, \bold{83(5)},1372--1385.

 Dambacher, J. M., et al. (2003) \href{http://www.journals.uchicago.edu/doi/full/10.1086/367590}{Qualitative stability and ambiguity in model ecosystems}. \emph{The American Naturalist}, \bold{161(6)},876--888.

 Dambacher, J. M., R. Levins and P. A. Rossignol. (2005) \href{https://doi.org/10.1016/j.mbs.2005.06.001}{Life expectancy change in perturbed communities: Derivation and qualitative analysis}. \emph{Mathematical Biosciences} \bold{197},1--14.

 Levins, R. (1974) \href{http://faculty.dbmi.pitt.edu/cosbbi/ibric2016/Levins-1974-Annals_of_the_New_York_Academy_of_Sciences.pdf}{The qualitative analysis of partially specified systems}. \emph{Annals of the New York Academy of Sciences}, \bold{231},123--138.

 Puccia, C. J. and Levins, R. (1986) \href{http://www.hup.harvard.edu/catalog.php?isbn=9780674435070}{\emph{Qualitative Modeling of Complex Systems: An Introduction to Loop Analysis and Time Averaging}}. Cambridge: Harvard University Press.
}
\keyword{ graphs }