\name{submatrix}
\docType{data}
\alias{submatrix}
\title{A Five-Variable Sub-Matrix of the Last Five Variables of Zavaleta's Matrix}
\description{
This dataset represents a subset of the causal relationships defining Zavaleta's five-variable system. Specifically, it contains the last five variables of the dataset \code{cm.zavaleta}.
}
\usage{data(submatrix)}
\format{A 5 by 5 (-1, 0, 1) matrix.  The matrix is interpreted as the jth variable has the direct causal effect a[i,j] on i.}
\references{
  Puccia, C. J. and Levins, R. (1986) \emph{Qualitative Modeling of Complex Systems: An Introduction to Loop Analysis and Time Averaging}. Cambridge: Harvard University Press.

Dambacher, J. M., Li, H. M., and Rossignol, P. A. (2002) Relevance of community structure in assessing indeterminacy of ecological predictions. \emph{Ecology}, \bold{83(5)},1372-1385
}
\keyword{datasets}
