\name{cem.levins}
\docType{data}
\alias{cem.levins}
\title{Community Effect Matrix For A Four-Variable Trophic Model}
\description{
This dataset contains the computed community effect matrix of the model in Figure 2 from Levins' 1996 paper (see references).
}
\usage{data(cem.levins)}
\format{A 4 by 4 (-1, 0, 1) matrix.  The matrix is interpreted as the ith parameter has the indirect causal effect [i,j] on parameter j after the system stabilizes in response to a perturbation.}
\references{
  Puccia, C. J. and Levins, R. (1986) \emph{Qualitative Modeling of Complex Systems: An Introduction to Loop Analysis and Time Averaging}. Cambridge: Harvard University Press.

Levins, R. and Schultz, B. B. (1996) Effects of density dependence, feedback and environmental sensitivity on correlations among predators, prey and plant resources: Models and practical implications. \emph{Journal of Animal Ecology}, \bold{65(6)},802-812
}
\keyword{datasets}
