\name{community.matrix}
\docType{data}
\alias{community.matrix}
\title{Simmulated Five-Parameter Data Set}
\description{
This dataset represents the causal relationships defining a five-parameter system.
}
\usage{data(community.matrix)}
\format{A 5 by 5 (-1, 0, 1) matrix with each element having at least one parent and at least one child parameter (node).  The matrix is interpreted as the ith parameter has the direct causal effect [i,j] on parameter j.}
\references{
  Puccia, C. J. and Levins, R. (1986) \emph{Qualitative Modeling of Complex Systems: An Introduction to Loop Analysis and Time Averaging}. Cambridge: Harvard University Press.
}
\keyword{datasets}
