\name{submatrix}
\docType{data}
\alias{submatrix}
\title{A Five-Variable Sub-Matrix of the First Five Variables of Dambacher's Matrix}
\description{
This dataset represents a subset of the causal relationships defining Dambacher's five-variable system. Specifically, it contains the first five variables of the dataset \code{cm.dambacher}.
}
\usage{data(submatrix)}
\format{A 5 by 5 (-1, 0, 1) matrix.  The matrix is interpreted as the jth variable has the direct causal effect a[i,j] on i.}
\references{
 Dambacher, J. M. and Li, H. W. and Rossignol, P. A. (2002) Relevance of community structure in assessing indeterminacy of ecological predictions. \emph{Ecology}, \bold{83(5)},1372--1385. <\doi{10.2307/3071950}>.

 Puccia, C. J. and Levins, R. (1986) \emph{Qualitative Modeling of Complex Systems: An Introduction to Loop Analysis and Time Averaging}. Cambridge: Harvard University Press. <\doi{10.4159/harvard.9780674435070}>.
}
\keyword{datasets}
