\name{submatrix}
\docType{data}
\alias{submatrix}
\title{A Five-Variable Sub-Matrix of the First Five Variables of Dambacher's Matrix}
\description{
This dataset represents a subset of the causal relationships defining Dambacher's five-variable system. Specifically, it contains the first five variables of the dataset \code{cm.dambacher}.
}
\usage{data(submatrix)}
\format{A 5 by 5 (-1, 0, 1) matrix.  The matrix is interpreted as the jth variable has the direct causal effect a[i,j] on i.}
\references{
 Dambacher, J. M. and Li, H. W. and Rossignol, P. A. (2002) \href{https://www.jstor.org/stable/3071950?seq=1#page_scan_tab_contents}{Relevance of community structure in assessing indeterminacy of ecological predictions}. \emph{Ecology}, \bold{83(5)},1372--1385.

 Puccia, C. J. and Levins, R. (1986) \href{http://www.hup.harvard.edu/catalog.php?isbn=9780674435070}{\emph{Qualitative Modeling of Complex Systems: An Introduction to Loop Analysis and Time Averaging}}. Cambridge: Harvard University Press.
}
\keyword{datasets}
