\name{submatrix}
\docType{data}
\alias{submatrix}
\title{A Three-Variable Sub-Matrix of a Simulated Five-Parameter Data Set}
\description{
This dataset represents a subset of the causal relationships defining a five-variable system.  Specifically, it contains the second, third and fourth variable of the dataset \code{community.matrix}.
}
\usage{data(submatrix)}
\format{A 3 by 3 (-1, 0, 1) matrix.  The matrix is interpreted as the jth variable has the direct causal effect a[i,j] on i.}
\references{
  Puccia, C. J. and Levins, R. (1986) \emph{Qualitative Modeling of Complex Systems: An Introduction to Loop Analysis and Time Averaging}. Cambridge: Harvard University Press.
}
\keyword{datasets}
