\name{submatrix}
\docType{data}
\alias{submatrix}
\title{A Three-Parameter Sub-Matrix of a Simulated Five-Parameter Data Set}
\description{
This dataset represents a subset of the causal relationships defining a five-parameter system.  Specifically, it contains the second, third and fourth parameters of the dataset \code{community.matrix}.
}
\usage{data(submatrix)}
\format{A 3 by 3 (-1, 0, 1) matrix.  The matrix is interpreted as the ith parameter has the direct causal effect [i,j] on parameter j.}
\references{
  Puccia, C. J. and Levins, R. (1986) \emph{Qualitative Modeling of Complex Systems: An Introduction to Loop Analysis and Time Averaging}. Cambridge: Harvard University Press.
}
\keyword{datasets}
